backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
SQLAlchemy>=2.0.29
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
groq>=0.5.0
python-dotenv>=1.0.1
